+++ src/pages/modify/index.tsx
@@ -1,8 +1,8 @@
 import styled from "styled-components";
 import { themedPalette } from "@/styles/theme";
 import ModifyImg from "@/components/modify/Img";
-import Input from "@/components/common/input";
-import ModifyButton from "@/components/modify/button";
+import Input from "@/components/common/Input";
+import ModifyButton from "@/components/modify/Button";
 import { memberApi } from "@/api/memberApi";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
